try to break down caching file into pieces
